renamed C file to math functions and added multiplication and division function
